Release 2.32. Fixed a self-introduced FC bug which happens if you choose a contact picture by using the Gallery/photo picker, and cancel it halfway. Fixed an AOSP bug with "Display Groups" in Contacts always having an index off. See comments in github code for more information.

+++ src/com/android/contacts/EditContactActivity.java
@@ -613,9 +613,11 @@ public final class EditContactActivity extends Activity implements View.OnClickL
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (resultCode != RESULT_OK) {
             //Delete temp file if it exists
-            File f = new File(mImageCaptureUri.getPath());            
-            if (f.exists()) {
-                f.delete();
+            if (mImageCaptureUri != null) {
+                File f = new File(mImageCaptureUri.getPath());            
+                if (f.exists()) {
+                    f.delete();
+                }
             }
             return;
         }
